chore: amend command typing
